# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Network Asset Discovery & NAC Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discovery/jobs:
    get:
      operationId: listDiscoveryJobs
      tags: [discovery]
      summary: List discovery jobs
      responses:
        "200":
          description: Discovery jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveryJob"
    post:
      operationId: createDiscoveryJob
      tags: [discovery]
      summary: Create and trigger a discovery job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscoveryJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryJob"
  /discovery/jobs/{id}:
    get:
      operationId: getDiscoveryJob
      tags: [discovery]
      summary: Get discovery job status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryJob"
  /discovery/sources:
    get:
      operationId: listDiscoverySources
      tags: [discovery]
      summary: List configured discovery sources
      responses:
        "200":
          description: Sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoverySource"
    post:
      operationId: createDiscoverySource
      tags: [discovery]
      summary: Add a discovery source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscoverySourceRequest"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySource"
  /discovery/sources/{id}:
    put:
      operationId: updateDiscoverySource
      tags: [discovery]
      summary: Update a discovery source
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscoverySourceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySource"
    delete:
      operationId: deleteDiscoverySource
      tags: [discovery]
      summary: Delete discovery source
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───────────────────────────── VLANs ─────────────────────────────
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all sites
      responses:
        "200":
          description: Site list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    put:
      operationId: updateSite
      tags: [sites]
      summary: Update site
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete site
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───────────────────────────── RADIUS ─────────────────────────────
  /radius/clients:
    get:
      operationId: listRadiusClients
      tags: [radius]
      summary: List RADIUS clients
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RadiusClient"
    post:
      operationId: createRadiusClient
      tags: [radius]
      summary: Add RADIUS client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRadiusClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiusClient"
  /radius/clients/{id}:
    put:
      operationId: updateRadiusClient
      tags: [radius]
      summary: Update RADIUS client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRadiusClientRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiusClient"
    delete:
      operationId: deleteRadiusClient
      tags: [radius]
      summary: Delete RADIUS client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /radius/groups:
    get:
      operationId: listRadiusGroups
      tags: [radius]
      summary: List RADIUS groups
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RadiusGroup"
    post:
      operationId: createRadiusGroup
      tags: [radius]
      summary: Create RADIUS group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRadiusGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiusGroup"
  /radius/groups/{id}:
    put:
      operationId: updateRadiusGroup
      tags: [radius]
      summary: Update RADIUS group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRadiusGroupRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiusGroup"
    delete:
      operationId: deleteRadiusGroup
      tags: [radius]
      summary: Delete RADIUS group
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /radius/sync:
    post:
      operationId: syncToRadius
      tags: [radius]
      summary: Sync approved devices to FreeRADIUS
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiusSyncInput"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List NAC policies
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NacPolicy"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create NAC policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNacPolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NacPolicy"
  /policies/{id}:
    put:
      operationId: updatePolicy
      tags: [policies]
      summary: Update NAC policy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNacPolicyRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NacPolicy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: Delete NAC policy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ───────────────────────────── ALERTS ─────────────────────────────
...
components:
  schemas:
...
    RadiusSyncInput:
      type: object
      properties:
        deviceIds:
          type: array
          items:
            type: integer
...
    Vlan:
      type: object
      required: [id, vlanId, name]
      properties:
        id:
          type: integer
        vlanId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [production, quarantine, guest, management]
          nullable: true
        isQuarantine:
          type: boolean
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Site:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [site, branch, building, floor, rack]
        deviceCount:
          type: integer
        switchCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateSiteRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        parentId:
          type: integer
        type:
          type: string
          enum: [site, branch, building, floor, rack]
    DiscoveryJob:
      type: object
      required: [id, type, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [snmp, dhcp, dns, ad, manual]
        status:
          type: string
          enum: [pending, running, completed, failed]
        sourceId:
          type: integer
          nullable: true
        sourceName:
          type: string
          nullable: true
        devicesFound:
          type: integer
        devicesNew:
          type: integer
        devicesUpdated:
          type: integer
        errorMessage:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateDiscoveryJobRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [snmp, dhcp, dns, ad, manual]
        sourceId:
          type: integer
    DiscoverySource:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [snmp, dhcp_windows, dhcp_isc, dns_windows, dns_bind, active_directory]
        host:
          type: string
        port:
          type: integer
          nullable: true
        enabled:
          type: boolean
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        lastRunStatus:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateDiscoverySourceRequest:
      type: object
      required: [name, type, host]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [snmp, dhcp_windows, dhcp_isc, dns_windows, dns_bind, active_directory]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        community:
          type: string
        enabled:
          type: boolean
    RadiusClient:
      type: object
      required: [id, name, ipAddress]
      properties:
        id:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        secret:
          type: string
        nasType:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateRadiusClientRequest:
      type: object
      required: [name, ipAddress, secret]
      properties:
        name:
          type: string
        ipAddress:
          type: string
        secret:
          type: string
        nasType:
          type: string
        description:
          type: string
    RadiusGroup:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        vlanId:
          type: integer
          nullable: true
        vlan:
          $ref: "#/components/schemas/Vlan"
          nullable: true
        attribute:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateRadiusGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        vlanId:
          type: integer
        attribute:
          type: string
        value:
          type: string
        description:
          type: string
    SyncResult:
      type: object
      required: [synced, failed]
      properties:
        synced:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    NacPolicy:
      type: object
      required: [id, name, condition, action, priority]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        condition:
          type: string
          enum: [known_device, unknown_device, rejected_device, port_change, duplicate_mac, duplicate_ip]
        action:
          type: string
          enum: [assign_vlan, deny_access, quarantine, alert, approve]
        priority:
          type: integer
        vlanId:
          type: integer
          nullable: true
        vlan:
          $ref: "#/components/schemas/Vlan"
          nullable: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateNacPolicyRequest:
      type: object
      required: [name, condition, action, priority]
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          type: string
          enum: [known_device, unknown_device, rejected_device, port_change, duplicate_mac, duplicate_ip]
        action:
          type: string
          enum: [assign_vlan, deny_access, quarantine, alert, approve]
        priority:
          type: integer
        vlanId:
          type: integer
        enabled:
          type: boolean
...