# lib/api-zod/src/generated/types/discoverySource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Asset Discovery & NAC Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoverySourceType } from './discoverySourceType';

export interface DiscoverySource {
  id: number;
  name: string;
  type: DiscoverySourceType;
  host?: string;
  port?: number | null;
  enabled?: boolean;
  lastRunAt?: Date | null;
  lastRunStatus?: string | null;
  createdAt?: Date;
}